browser/components/aboutlogins/tests/chrome/test_menu_button.html

<!DOCTYPE HTML>
<html>
<!--
Test the menu-button component
-->
<head>
  <meta charset="utf-8">
  <title>Test the menu-button component</title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>
  <script type="module" src="chrome://browser/content/aboutlogins/components/menu-button.js"></script>
  <script src="aboutlogins_common.js"></script>

  <link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
  <p id="display">
  </p>
<div id="content" style="display: none">
  <iframe id="templateFrame" src="chrome://browser/content/aboutlogins/aboutLogins.html"
          sandbox="allow-same-origin"></iframe>
</div>
<pre id="test">
</pre>
<script>
/** Test the menu-button component **/

let gMenuButton;
add_task(async function setup() {
  let templateFrame = document.getElementById("templateFrame");
  let displayEl = document.getElementById("display");
  importDependencies(templateFrame, displayEl);

  gMenuButton = document.createElement("menu-button");
  displayEl.appendChild(gMenuButton);

  isnot(document.activeElement, gMenuButton, "menu-button should not be focused by default");
  while (document.activeElement != gMenuButton) {
    sendKey("TAB");
    await new Promise(resolve => requestAnimationFrame(resolve));
  }
});

add_task(async function test_menu_open_close() {
  is(document.activeElement, gMenuButton, "menu-button should be focused to start the test");

  let menu = gMenuButton.shadowRoot.querySelector(".menu");
  is(true, menu.hidden, "menu should be hidden before pressing 'space'");
  sendKey("SPACE");
  await new Promise(resolve => requestAnimationFrame(resolve));
  ok(!menu.hidden, "menu should be visible after pressing 'space'");

  sendKey("ESCAPE");
  await new Promise(resolve => requestAnimationFrame(resolve));
  ok(menu.hidden, "menu should be hidden after pressing 'escape'");
  is(gMenuButton.shadowRoot.activeElement, gMenuButton.shadowRoot.querySelector(".menu-button"),
    "the .menu-button should be focused after closing the menu via keyboard");

  sendKey("RETURN");
  await new Promise(resolve => requestAnimationFrame(resolve));
  ok(!menu.hidden, "menu should be visible after pressing 'return'");

  let firstVisibleItem = gMenuButton.shadowRoot.querySelector(".menuitem-button:not([hidden])");
  ok(!firstVisibleItem.matches(":focus"), "the first item should not be focused before tabbing to it");
  sendKey("TAB");
  await SimpleTest.promiseWaitForCondition(() => firstVisibleItem.matches(":focus"),
    "waiting for firstVisibleItem to get focus");
  ok(firstVisibleItem.matches(":focus"), "firstVisibleItem should be focused after tabbing to it");
  synthesizeKey("VK_TAB", { shiftKey: true });
  await SimpleTest.promiseWaitForCondition(() => !firstVisibleItem.matches(":focus"),
    "waiting for firstVisibleItem to lose focus");
  ok(!firstVisibleItem.matches(":focus"), "firstVisibleItem should lose focus after tabbing away from it");
  sendKey("TAB");
  await SimpleTest.promiseWaitForCondition(() => firstVisibleItem.matches(":focus"),
    "waiting for firstVisibleItem to get focus again");
  ok(firstVisibleItem.matches(":focus"), "firstVisibleItem should be focused after tabbing to it again");

  if (navigator.platform == "Win32") {
    // The Import menuitem is only visible on Windows, where we will need another Tab
    // press to get to the Preferences item.
    let preferencesItem = gMenuButton.shadowRoot.querySelector(".menuitem-preferences");
    sendKey("DOWN");
    await SimpleTest.promiseWaitForCondition(() => preferencesItem.matches(":focus"),
      "waiting for preferencesItem to gain focus");
    ok(preferencesItem.matches(":focus"), `.menuitem-preferences should be now be focused (DOWN)`);
    sendKey("UP");
    await SimpleTest.promiseWaitForCondition(() => !preferencesItem.matches(":focus"),
      `waiting for preferencesItem to lose focus (UP)`);
    ok(!preferencesItem.matches(":focus"), `.menuitem-preferences should lose focus after pressing up`);

    sendKey("TAB");
    await SimpleTest.promiseWaitForCondition(() => preferencesItem.matches(":focus"),
      "waiting for preferencesItem to get focus");
    ok(preferencesItem.matches(":focus"), ".menuitem-preferences should be focused after tabbing to it");
  }

  let openPreferencesEvent = null;
  ok(!menu.hidden, "menu should be visible before pressing 'space' on .menuitem-preferences");
  window.addEventListener("AboutLoginsOpenPreferences", event => openPreferencesEvent = event);
  sendKey("SPACE");
  ok(openPreferencesEvent, "AboutLoginsOpenPreferences event should be dispatched after pressing 'space' on .menuitem-preferences");
  ok(menu.hidden, "menu should be hidden after pressing 'space' on .menuitem-preferences");
});
</script>

</body>
</html>
